feat(useTitle-hook): add custom hook in outside folder, implement hook in views
